backend/train_requirements.txt
# Minimal deps for training only (no DB/API). Use with: source .venv/bin/activate && pip install -r train_requirements.txt
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
PyYAML
tqdm
matplotlib
scikit-learn
